services/pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Event-driven price check pipeline Lambdas"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "pydantic>=2.0",
    "structlog>=24.0",
    "boto3>=1.34",
    "google-generativeai>=0.8",
    "openai>=1.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
]
